Update to MonetaryUnit Stats.
Issue was 120 of 129 masternodes where displaying "NEW_START_REQUIRED". Which under normal cases means the masternodes are not receiving rewards. Based on information received on `bizzyb` form the muecoin slack we where notified that this is not the case on mue. Issue was resolved and stats are now showing correct data based on `bizzyb`.

bizzyb[7:51 AM]
we don't have payment enforcement on, so new start required MNs are earning fine